Bug fix for Cantherm
We wre not transforming our coordinates before projecting out the rotors,
apparently this was assumed by the Gaussian white paper formulas.

This result in the frequencies after prjecting to be 10-40 cm-1
lower than the HO frequencies which can result in an error in the
ZPE. This bug can have a big effect for low temperatures and in reactions
where the ZPE between the reactants and TS do not cancel each other out.

Thanks to @agvandeputte for identifying this and also providing the fix